# docs/swagger.yaml
definitions:
  github_com_maczh_mgin_models.Result-any:
    properties:
      data: {}
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
  github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysApi:
    properties:
      data:
        $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysApi'
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
  github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDept:
    properties:
      data:
        $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysDept'
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
...
  github_com_maczh_mgin_models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  github_com_maczh_mgin_models_sys.SysApi:
    properties:
      api_group:
        type: string
      api_path:
        type: string
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      description:
        type: string
      enabled:
        type: integer
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      need_auth:
        type: integer
      need_log:
        type: integer
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
    type: object
  github_com_maczh_mgin_models_sys.SysDept:
    properties:
      ancestors:
        type: string
      children:
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysDept'
        type: array
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      deptType:
        type: string
      id:
        type: integer
      leader:
        type: string
      mobile:
        type: string
      name:
        type: string
      parentId:
        type: integer
      parentName:
        type: string
      sort:
        type: integer
      status:
        type: integer
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
    type: object
...
  github_com_maczh_mgin_models_sys_request.CreateDeptReq:
    properties:
      ancestors:
        type: string
      leader:
        type: string
      mobile:
        type: string
      name:
        type: string
      parentId:
        type: integer
      sort:
        type: integer
      type:
        type: string
    required:
    - name
    type: object
...
  github_com_maczh_mgin_models_sys_request.DeleteByIdReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  github_com_maczh_mgin_models_sys_request.VerifyCaptchaReq:
    properties:
      code:
        type: string
      id:
        type: string
    required:
    - code
    - id
    type: object
...
  models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/v1/captcha/get:
    get:
      consumes:
      - application/json
      description: 获取图片验证码，返回验证码 ID 和 Base64 编码的图片
      parameters:
      - in: query
        name: height
        required: true
        type: integer
      - in: query
        name: length
        type: integer
      - description: 0:数字 1:字母 2:算术
        in: query
        name: type
        type: integer
      - in: query
        name: width
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回验证码信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取验证码失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取验证码
      tags:
      - 验证码
  /api/v1/captcha/verify:
    post:
      consumes:
      - application/json
      description: 验证用户输入的验证码是否正确
      parameters:
      - description: 验证验证码请求参数
        in: body
        name: verifyReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.VerifyCaptchaReq'
      produces:
      - application/json
      responses:
        "200":
          description: 验证码验证成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或验证码错误
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 验证验证码
      tags:
      - 验证码
...
  /api/v1/sys/dept/add:
    post:
      consumes:
      - application/json
      description: 创建一个新的部门信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 部门创建请求参数
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.CreateDeptReq'
      produces:
      - application/json
      responses:
        "200":
          description: 新增成功，返回新增的部门信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDept'
        "500":
          description: 参数绑定失败或新增部门失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 新增部门
      tags:
      - 系统部门
  /api/v1/sys/dept/del:
    post:
      consumes:
      - application/json
      description: 根据部门 ID 删除指定部门，软删除
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或删除部门失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除部门
      tags:
      - 系统部门
  /api/v1/sys/dept/get:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取指定部门的信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回部门信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDept'
        "500":
          description: 参数绑定失败或获取部门信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取部门信息
      tags:
      - 系统部门
  /api/v1/sys/dept/list:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取部门的分页列表信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回部门列表和总数
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取部门列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取部门列表
      tags:
      - 系统部门
  /api/v1/sys/dept/tree:
    get:
      consumes:
      - application/json
      description: 根据部门 ID 获取部门的树形结构信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回部门树信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取部门树失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取部门树
      tags:
      - 系统部门
  /api/v1/sys/dept/update:
    post:
      consumes:
      - application/json
      description: 更新指定部门的相关信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新后的部门信息
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysDept'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或更新部门信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 更新部门信息
      tags:
      - 系统部门
...
  /api/v1/sys_api/add:
    post:
      consumes:
      - application/json
      description: 创建一个新的API接口
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: API接口信息
        in: body
        name: api
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回创建的API接口信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysApi'
        "500":
          description: 请求参数错误或创建API接口失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 创建API接口
      tags:
      - 系统API
  /api/v1/sys_api/del:
    post:
      consumes:
      - application/json
      description: 根据ID软删除指定的API接口
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: API接口ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.DeleteByIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或删除API接口失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除API接口
      tags:
      - 系统API
  /api/v1/sys_api/get:
    get:
      consumes:
      - application/json
      description: 根据ID获取API接口信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回API接口信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysApi'
        "500":
          description: 请求参数错误或获取API接口失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取API接口
      tags:
      - 系统API
  /api/v1/sys_api/get/uri:
    get:
      consumes:
      - application/json
      description: 根据URI获取对应的API接口信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: uri
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回API接口信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysApi'
        "500":
          description: 请求参数错误或获取API接口失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 根据URI获取API接口
      tags:
      - 系统API
  /api/v1/sys_api/group:
    get:
      consumes:
      - application/json
      description: 按模块分组获取API接口列表, for 前端使用
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回API接口列表和分页信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.ResultPage'
        "500":
          description: 参数绑定失败或获取API接口列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 按模块分组获取API接口列表
      tags:
      - 系统API
  /api/v1/sys_api/list:
    get:
      consumes:
      - application/json
      description: 获取API接口的分页列表
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: group
        type: string
      - in: query
        name: needAuth
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回API接口列表和分页信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.ResultPage'
        "500":
          description: 参数绑定失败或获取API接口列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取API接口列表
      tags:
      - 系统API
  /api/v1/sys_api/update:
    post:
      consumes:
      - application/json
      description: 更新指定API接口的信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新后的API接口信息
        in: body
        name: api
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或更新API接口失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 更新API接口
      tags:
      - 系统API
...
swagger: "2.0"